# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel E-commerce Store API
  description: |
    API RESTful completa para o Laravel E-commerce Store.

    Esta API oferece endpoints para:
    - Gestão de produtos, categorias e marcas
    - Carrinho de compras e pedidos
    - Sistema de avaliações e comentários
    - Gestão de clientes e endereços
    - Cupons e promoções
    - Relatórios e analytics
    - Administração completa

    ## Autenticação

    A API utiliza Laravel Sanctum para autenticação. Para acessar endpoints protegidos:

    1. Faça login para obter um token
    2. Inclua o token no header: `Authorization: Bearer {token}`
    3. Para operações administrativas, o usuário deve ter privilégios de admin

    ## Paginação

    Todos os endpoints que retornam listas suportam paginação:

    ```json
    {
      "success": true,
      "data": [...],
      "pagination": {
        "current_page": 1,
        "per_page": 15,
        "total": 100,
        "last_page": 7,
        "from": 1,
        "to": 15
      }
    }
    ```

    ## Filtros e Ordenação

    Use query parameters para filtrar e ordenar:

    - `?sort=name&order=asc` - Ordenar por nome (asc/desc)
    - `?category_id=1` - Filtrar por categoria
    - `?min_price=100&max_price=1000` - Filtrar por preço
    - `?search=smartphone` - Buscar produtos
    - `?in_stock=1` - Apenas produtos em estoque
    - `?featured=1` - Apenas produtos em destaque

    ## Cache

    A API utiliza cache inteligente com headers informativos:

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKey: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de autenticação obtido via `/api/v1/auth/login`.

        Inclua no header: `Authorization: Bearer {token}`
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Chave de API para acesso administrativo.

        Obtenha em: `/api/v1/admin/api-keys`
  schemas:
...
    Review:
      type: object
      properties:
        id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 1
        customer_id:
          type: integer
          example: 1
        order_id:
          type: integer
          example: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Excelente produto! Recomendo muito."
        pros:
          type: string
          example: "Câmera excelente, bateria dura o dia todo"
        cons:
          type: string
          example: "Preço um pouco elevado"
        is_verified:
          type: boolean
          example: true
        is_approved:
          type: boolean
          example: true
        helpful_votes:
          type: integer
          example: 15
        total_votes:
          type: integer
          example: 20
        customer_name:
          type: string
          example: "João Silva"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erro de validação"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            name: ["O campo nome é obrigatório"]
            email: ["O campo email deve ser um email válido"]
        code:
          type: integer
          example: 422
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 15
        total:
          type: integer
          example: 100
        last_page:
          type: integer
          example: 7
        from:
          type: integer
          example: 1
        to:
          type: integer
          example: 15
        has_more_pages:
          type: boolean
          example: true
        next_page_url:
          type: string
          nullable: true
          example: "https://api.loja.com/api/v1/products?page=2"
        prev_page_url:
          type: string
          nullable: true
          example: null
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
          example: "Operação realizada com sucesso"
        pagination:
          $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
...
paths:
...
  /api/v1/reviews:
    get:
      tags:
        - Reviews
      summary: Listar avaliações
      description: Retorna avaliações com paginação e filtros.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Página atual
        - name: per_page
          in: query
          schema:
            type: integer
            default: 15
          description: Itens por página
        - name: product_id
          in: query
          schema:
            type: integer
          description: Filtrar por produto
        - name: rating
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
          description: Filtrar por avaliação
        - name: verified
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: Apenas avaliações verificadas
      responses:
        '200':
          description: Lista de avaliações
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Review'
    post:
      tags:
        - Reviews
      summary: Criar avaliação
      description: |
        Cria uma nova avaliação para um produto.

        **Requisitos:**
        - Usuário deve estar autenticado
        - Usuário deve ter comprado o produto
        - Produto deve existir e estar ativo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - rating
                - comment
              properties:
                product_id:
                  type: integer
                  example: 1
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
                comment:
                  type: string
                  minLength: 10
                  maxLength: 1000
                  example: "Excelente produto! Recomendo muito."
                pros:
                  type: string
                  example: "Câmera excelente, bateria dura o dia todo"
                cons:
                  type: string
                  example: "Preço um pouco elevado"
      responses:
        '201':
          description: Avaliação criada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Review'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...